Input patch.
An attempt to maintain full key state between Xinput and the core X keyboard API
(principally necessitated by the addition of scroll_* options).

NOTE: Maintaining additional state is suboptimal, relies on certain assumptions
about the Xlib API and introduces the possibility of new bugs.

+++ src/hints.c
@@ -97,7 +97,7 @@ static size_t read_hints(const char *path,
 		line[len-1] = '\0';
 
 		assert(i < MAX_HINTS);
-		assert(len < sizeof hints[0].label);
+		assert(len < (ssize_t)sizeof hints[0].label);
 
 		strcpy(hints[i++].label, line);
 		free(line);
